refactor: snake case component names; remove barrel fiel
