--- src/common/components/CustomIcon/index.ts
@@ -1,1 +1,0 @@
-export * from './CustomIcon';
