+++ src/common/components/custom-icon/README.md
